refactor: IssueSidebar 컴포넌트 생성 및 폴더구조 수정
- issue/_components 폴더 안에 컴포넌트  및 스타일 파일 이동
- layout에는 IssueSidebar 컴포넌트만 import하여 사용
