+++ src/app/issue/layout.tsx
@@ -1,36 +1,10 @@
-import Sidebar from '@/components/Sidebar/Sidebar';
-import SidebarItem from '@/components/Sidebar/SidebarItem';
-import * as S from '@/components/Sidebar/styles';
-import IssueGraphLink from './IssueGraphLink';
-import NewIssueButton from './NewIssueButton';
-
-const ISSUE_LIST = [
-  { title: 'new issue', href: '#', status: 'open' },
-  { title: 'voting issue', href: '#', status: 'voting' },
-  { title: 'closed issue', href: '#', status: 'closed' },
-] as const;
+import IssueSidebar from './_components/IssueSidebar';
 
 export default function layout() {
   return (
     <>
       <header>header</header>
-      <Sidebar>
-        <IssueGraphLink />
-        <S.SidebarTitle>
-          ISSUE LIST
-          <NewIssueButton />
-        </S.SidebarTitle>
-        <S.SidebarList>
-          {ISSUE_LIST.map((issue) => (
-            <SidebarItem
-              key={issue.title}
-              title={issue.title}
-              href={issue.href}
-              status={issue.status}
-            />
-          ))}
-        </S.SidebarList>
-      </Sidebar>
+      <IssueSidebar />
     </>
   );
 }
